odroid: Add support for the ODROID-X board variant
ODROID-X uses a slightly older revision of the same base board
as the ODROID-X2. But the CPU module in ODROID-X uses an older
1.4GHz revision of Exynos4412 SoC and less RAM (1GiB instead
of 2GiB).

The current U-Boot code deadlocks on ODROID-X when probing the RAM
size via get_ram_size() function. Reducing CONFIG_NR_DRAM_BANKS
from 8 to 4 can fix this problem. But this constant is used in
a lot of places in U-Boot, while SDRAM_BANK_SIZE can be easily
replaced with the code which relies on runtime detection of the
board type. So we change 4 or 8 banks of 256MiB with just one
fake bank of 2GiB or 1GiB. The runtime detection check tries to
read the PRO_ID register in the hope that it might be different
between Exynos4412 and Exynos4412 Prime and enough to detect
the difference between X and X2 board variants.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>