HACK: usb: exynos: Explicitly call board_usb_init() function
During DM conversion of the Exynos USB drivers, a call to
the board_usb_init() function was lost. As a result, USB does
not initialize correctly on ODROID boards and the (USB based)
Ethernet does not work too. More detailed explanations can
be found in the U-Boot mailing list:

    https://lists.denx.de/pipermail/u-boot/2015-October/231061.html

This patch is a crude hack, but it fixes the Ethernet support
on ODROID-X and maybe other similar ODROID boards.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>